Mutates existing work into a working example
Finishing features are added to demonstrate a singularity image can be
made within a docker container.

A shell script replaces the (perhaps better) makefile approach to tune
the demo for personal preferences/capabilities.

Usage
-----

Singularity definition files and the output containers are all passed to
the container through the `shared` folder and paths given to the
spack-docker container (with `singularity` command as the entry-point)
access the path through the volume `/shared`. See `build.sh` for
examples running.